vision_agent/requirements.txt
Flask==3.0.1
gunicorn==22.0.0

requests==2.32.0
structlog==24.1.0
google-cloud-firestore
google-auth==2.3.2

pg8000
sqlalchemy
Flask-SQLAlchemy

google-cloud-texttospeech

firebase_admin

openai
